Adiciona novos endpoints para listar avaliações completas e obter estatísticas de voluntários, além de ajustes nos DTOs e entidades para suportar tipos de dados corretos (double para Latitude e Longitude).

+++ AcessibilidadeWebAPI/Entidades/Local.cs
@@ -7,8 +7,8 @@
             AvaliacaoLocals = new HashSet<AvaliacaoLocal>();
         }
         public int IdLocal { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public string Descricao { get; set; }
         public int AvaliacaoAcessibilidade { get; set; }
         public virtual ICollection<AvaliacaoLocal> AvaliacaoLocals { get; set; }
